split VerifiableCalledTests
