+++ Tests/MockSwiftTests/Core/Then/Verifiables/VerifiableCalledTests.swift
@@ -143,37 +143,6 @@ class VerifiableCalledTests: XCTestCase {
         XCTAssertEqual(failureRecorder.recordFailureReceived.count, 0)
     }
 
-    // MARK: - neverCalled
-
-    func test_neverCalled_should_recordFailure_when_recordedCallsReturnCalls() {
-        // Given
-        callRegister.recordedCallsReturn = [.stub(), .stub()]
-        predicate.description = "description"
-
-        // When
-        verifiable.neverCalled(file: "file", line: 42)
-
-        // Then
-        callRegisterCalled(with: functionIdentifier, and: predicate)
-
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 1)
-        let (message, file, line) = failureRecorder.recordFailureReceived[0]
-        XCTAssertEqual(message, "function(arg: description) -> Int " +
-                        "is expected to be called 0 time(s) but is called 2 time(s).")
-        XCTAssertEqual("\(file) \(line)", "file 42")
-    }
-
-    func test_neverCalled_should_notRecordFailure_when_recordedCallsReturnNoCalls() {
-        // Given
-        callRegister.recordedCallsReturn = []
-
-        // When
-        verifiable.neverCalled()
-
-        // Then
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 0)
-    }
-
     // MARK: - calledOnce
 
     func test_calledOnce_should_recordFailure_when_recordedCallsReturnNoCalls() {
@@ -320,99 +289,6 @@ class VerifiableCalledTests: XCTestCase {
         XCTAssertEqual(callAssertion?.functionIdentifier, functionIdentifier)
         XCTAssertEqual(callAssertion?.parametersPredicates.map { $0.description }, [predicate.description])
         XCTAssertEqual(callAssertion?.calls, [expectedCall1, expectedCall2])
-        XCTAssertTrue(callAssertion?.previous as AnyObject === assertion)
-    }
-
-    // MARK: - neverCalled after Assertion
-
-    func test_neverCalledAfterAssertion_should_notRecordFailure_when_recordedCallsReturnEmpty() {
-        // Given
-        callRegister.recordedCallsReturn = []
-        predicate.description = "description"
-        let assertion = AssertionMock()
-        assertion.descriptionReturn = "assertion description"
-        assertion.isValidReturn = true
-        assertion.firstValidTimeReturn = 0
-
-        // When
-        verifiable.neverCalled(after: assertion, file: "file", line: 42)
-
-        // Then
-        callRegisterCalled(with: functionIdentifier, and: predicate)
-
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 0)
-    }
-
-    func test_neverCalledAfterAssertion_should_recordFailure_when_recordedCallsReturnCall() {
-        // Given
-        callRegister.recordedCallsReturn = [.stub(time: 0), .stub(time: 2)]
-        predicate.description = "description"
-        let assertion = AssertionMock()
-        assertion.descriptionReturn = "assertion description"
-        assertion.isValidReturn = true
-        assertion.firstValidTimeReturn = 1
-
-        // When
-        verifiable.neverCalled(after: assertion, file: "file", line: 42)
-
-        // Then
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 1)
-        let (message, file, line) = failureRecorder.recordFailureReceived[0]
-        XCTAssertEqual(message, "function(arg: description) -> Int is expected to be called 0 time(s) " +
-                        "but is called 1 time(s) " +
-                        "after assertion description.")
-        XCTAssertEqual("\(file) \(line)", "file 42")
-    }
-
-    func test_neverCalledAfterAssertion_should_recordFailure_when_previousAssertionIsNotValid() {
-        // Given
-        callRegister.recordedCallsReturn = []
-        let assertion = AssertionMock()
-        assertion.descriptionReturn = "assertion description"
-        assertion.isValidReturn = false
-
-        // When
-        verifiable.neverCalled(after: assertion, file: "file", line: 42)
-
-        // Then
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 1)
-        let (message, file, line) = failureRecorder.recordFailureReceived[0]
-        XCTAssertEqual(message, "assertion description")
-        XCTAssertEqual("\(file) \(line)", "file 42")
-    }
-
-    func test_neverCalledAfterAssertion_should_notRecordFailure_when_recordedCallsReturnNoCall() {
-        // Given
-        callRegister.recordedCallsReturn = [.stub(time: 0), .stub(time: 1)]
-        let assertion = AssertionMock()
-        assertion.isValidReturn = true
-        assertion.firstValidTimeReturn = 2
-
-        // When
-        verifiable.neverCalled(after: assertion, file: "file", line: 42)
-
-        // Then
-        XCTAssertEqual(failureRecorder.recordFailureReceived.count, 0)
-    }
-
-    func test_neverCalledAfterAssertion_should_returnCorrectAssertion() {
-        // Given
-        callRegister.recordedCallsReturn = [.stub(time: 0)]
-        predicate.description = "description"
-        let assertion = AssertionMock()
-        assertion.isValidReturn = true
-        assertion.firstValidTimeReturn = 1
-
-        // When
-        let result = verifiable.neverCalled(after: assertion, file: "file", line: 42)
-
-        // Then
-        XCTAssertTrue(result is CallAssertion)
-        let callAssertion = result as? CallAssertion
-        XCTAssertEqual(callAssertion?.times.description, (==0).description)
-        XCTAssertEqual(callAssertion?.functionIdentifier, functionIdentifier)
-        XCTAssertEqual(callAssertion?.parametersPredicates.map { $0.description }, [predicate.description])
-        XCTAssertEqual(callAssertion?.calls.count, 0)
         XCTAssertTrue(callAssertion?.previous as AnyObject === assertion)
     }
 }
